% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateFact.R
\name{generateFact}
\alias{generateFact}
\title{Print pineapple facts}
\usage{
generateFact(number = sample(1:16, 1))
}
\arguments{
\item{number}{Fact number to be printed from 1 to 16. Defaults to be random.}
}
\description{
Prints on the console a pineapple fact from a list of 16. The selected fact can be chosen or can be random.
}
